requirements.txt
imageai==3.0.3                   
numpy==1.23.5
opencv-python==4.7.0.72
Pillow==9.5.0
telebot==0.0.5
torch==2.0.0
torchvision==0.15.1
